+++ TSS_Demo/examples/FRDMKLXX_DEMO/src/project1/mma8451.h
@@ -24,7 +24,7 @@ void read_full_xyz(void);
 void read_xyz(void);
 void convert_xyz_to_roll_pitch(void);
 
-extern float roll, pitch;
+extern uint8_t roll, pitch;
 extern int16_t acc_X, acc_Y, acc_Z;
 
 #endif
